+++ src/main/java/focik/net/progasoffice/addresses/domain/exceptions/AddressAlreadyExistsException.java
@@ -4,7 +4,7 @@ package focik.net.progasoffice.addresses.domain.exceptions;
 import focik.net.progasoffice.common.exceptions.ObjectAlreadyExistException;
 
 public class AddressAlreadyExistsException extends ObjectAlreadyExistException {
-    public AddressAlreadyExistsException(Integer id) {
+    public AddressAlreadyExistsException(Long id) {
         super("Address with id: " + id + " already exists.");
     }
 }
