+++ src/main/java/focik/net/progasoffice/addresses/domain/exceptions/AddressNotFoundException.java
@@ -4,7 +4,7 @@ package focik.net.progasoffice.addresses.domain.exceptions;
 import focik.net.progasoffice.common.exceptions.ObjectNotFoundException;
 
 public class AddressNotFoundException extends ObjectNotFoundException {
-    public AddressNotFoundException(Integer id) {
+    public AddressNotFoundException(Long id) {
         super("Address with id = " + id + " does not exist");
     }
 }
