+++ src/main/java/focik/net/progasoffice/addresses/domain/port/AddressRepository.java
@@ -1,8 +1,7 @@
 package focik.net.progasoffice.addresses.domain.port;
 
-import focik.net.progasoffice.addresses.domain.Address;
 import org.springframework.stereotype.Component;
-
+import focik.net.progasoffice.addresses.domain.Address;
 import java.util.Optional;
 
 @Component
@@ -10,7 +9,7 @@ public interface AddressRepository {
 
     Address save(Address address);
 
-    Optional<Address> findById(Integer id);
+    Optional<Address> findById(Long id);
 
-    void deleteAddress(Integer id);
+    void deleteAddress(Long id);
 }
