+++ src/main/java/focik/net/progasoffice/addresses/infrastructure/jpa/IAddressDtoRepository.java
@@ -4,5 +4,5 @@ package focik.net.progasoffice.addresses.infrastructure.jpa;
 import focik.net.progasoffice.addresses.infrastructure.dto.AddressDbDto;
 import org.springframework.data.repository.CrudRepository;
 
-interface IAddressDtoRepository extends CrudRepository<AddressDbDto, Integer> {
+interface IAddressDtoRepository extends CrudRepository<AddressDbDto, Long> {
 }
